requirements.txt
openai
numpy
python-dotenv
tiktoken
pathway
